Crate tuxedo_template_wallet

A simple CLI wallet. For now it is a toy just to start testing things out.

Modules

amoeba 🔒

Toy off-chain process to create an amoeba and perform mitosis on it

money 🔒

Wallet features related to spending money and checking balances.

Structs

Cli 🔒

The wallet’s main CLI struct

SpendArgs

Enums

Command 🔒

The tasks supported by the wallet

Constants

DEFAULT_ENDPOINT 🔒

The default RPC endpoint for the wallet to connect to

SHAWN_PHRASE 🔒

A default seed phrase for signing inputs when none is provided Corresponds to the default pubkey.

SHAWN_PUB_KEY 🔒

A default pubkey for receiving outputs when none is provided Corresponds to the default seed phrase

Statics

DEFAULT_VALUE 🔒

DEFAULT_VALUE 🔒

DEFAULT_VALUE 🔒

DEFAULT_VALUE 🔒

DEFAULT_VALUE 🔒

DEFAULT_VALUE 🔒

Functions

fetch_storage

Fetch an output from chain storage given an OutputRef

main 🔒

Opaque Types

🔒