Module tuxedo_template_wallet::keystore

Wallet’s local keystore.

This is a thin wrapper around sc-cli for use in tuxedo wallets

Constants

KEY_TYPE 🔒

A KeyTypeId to use in the keystore for Tuxedo transactions. We’ll use this everywhere until it becomes clear that there is a reason to use multiple of them

SHAWN_PHRASE

A default seed phrase for signing inputs when none is provided Corresponds to the default pubkey.

SHAWN_PUB_KEY

The public key corresponding to the default seed above.

Functions

generate_key

Generate a new key from system entropy and insert it into the keystore, optionally protected by a password.

get_keys

has_key

Check whether a specific key is in the keystore

insert_default_key_for_this_session

Insert the example “Shawn” key into the keystore for the current session only. This is convenient for ux when developing.

insert_key

Insert the private key associated with the given seed into the keystore for later use.

remove_key

Caution. Removes key from keystore. Call with care.

sign_with

Sign a given message with the private key that corresponds to the given public key.