Module tuxedo_template_wallet::parachain

Parachain compatibility for the template wallet.

The wallet is not intended to support a wide variety of chains, but it is able to support both the sovereign and parachain template nodes. There are a few types necessary to make this work.

We don’t want the wallet to depend on the parachain runtime which has a huge dependency graph itself. So a few types are duplicated here.

Enums

ParachainConstraintChecker

We don’t want the wallet to depend on the huge parachain codebase, So we just recreate this one little type here.