Crate tuxedo_template_wallet

A simple CLI wallet. For now it is a toy just to start testing things out.

Modules

amoeba 🔒

Toy off-chain process to create an amoeba and perform mitosis on it

cli 🔒

Tuxedo Template Wallet’s Command Line Interface.

keystore 🔒

Wallet’s local keystore.

money 🔒

Wallet features related to spending money and checking balances.

output_filter 🔒

rpc 🔒

Strongly typed helper functions for communicating with the Node’s RPC endpoint.

sync 🔒

This module is responsible for maintaining the wallet’s local database of blocks and owned UTXOs to the canonical database reported by the node.

Constants

DEFAULT_ENDPOINT 🔒

The default RPC endpoint for the wallet to connect to

Functions

default_data_path 🔒

Generate the platform-specific default data path for the wallet

fetch_storage

Fetch an output from chain storage given an OutputRef

h256_from_string 🔒

Parse a string into an H256 that represents a public key

main 🔒

output_ref_from_string 🔒

Parse an output ref from a string

pretty_print_verifier

Utility to pretty print an outer verifier

strip_0x_prefix 🔒

Takes a string and checks for a 0x prefix. Returns a string without a 0x prefix.